requirements.txt
openai==1.6.1
streamlit==1.29.0
langchain==0.0.352
stability-sdk 
pillow==9.3.0